requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2
django-filter==25.2
djangorestframework==3.16.1
drf-yasg==1.21.11
geographiclib==2.1
geopandas==1.1.1
geopy==2.4.1
git-filter-repo==2.47.0
idna==3.11
inflection==0.5.1
Markdown==3.10
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pyogrio==0.12.1
pyproj==3.7.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
shapely==2.1.2
six==1.17.0
sqlparse==0.5.4
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.1
